Use USHRT_MAX instead if INT_MAX since INT_MAX is too big for Qt
git-svn-id: https://qlc.svn.sourceforge.net/svnroot/qlc/trunk@2859 0db28fee-5c31-0410-96ea-e69cce70e049

+++ ui/src/grandmasterslider.cpp
@@ -45,7 +45,7 @@ GrandMasterSlider::GrandMasterSlider(QWidget* parent, OutputMap* outputMap, Inpu
     setSizePolicy(QSizePolicy::Maximum, QSizePolicy::MinimumExpanding);
 
     setMinimumSize(QSize(40, 100));
-    setMaximumSize(QSize(40, INT_MAX));
+    setMaximumSize(QSize(40, USHRT_MAX));
 
     new QVBoxLayout(this);
     layout()->setMargin(2);
